backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2 